+++ src/components/map/map-layer.tsx
@@ -7,21 +7,36 @@ interface MarkerLayerProps {
   id: string;
   annotationType: string;
   source: string;
+  visibility: boolean;
 }
 
 const MapLayer: React.FC<MarkerLayerProps> = ({
   id,
   annotationType,
   source,
+  visibility,
 }) => {
   const mapContext = useContext(MapContext); // Get the map context
   const map: Map | null = mapContext?.map || null;
 
   useEffect(() => {
+    if (!map) return;
+
+    const layer = map.getLayer(annotationType);
+
+    if (layer) {
+      map.setLayoutProperty(
+        layer.id,
+        "visibility",
+        visibility ? "visible" : "none",
+      );
+    }
+  }, [map, annotationType, visibility]);
+
+  useEffect(() => {
     async function addIconImageToMap() {
       // Get icon image file path
       const iconImagePath = getIconImagePath(annotationType);
-
       // Create a new Image element
       const img: HTMLImageElement = new Image();
       // Set the source of the image to the file path
@@ -40,12 +55,14 @@ const MapLayer: React.FC<MarkerLayerProps> = ({
               if (image && !map.hasImage(annotationType)) {
                 map.addImage(annotationType, image); // Use markerType as the icon uuid
               }
+
               if (!map.getLayer(id)) {
                 map.addLayer({
                   id: id,
                   type: "symbol",
                   source: source,
                   layout: {
+                    visibility: "visible",
                     "icon-image": annotationType,
                     "icon-size": 1.0,
                     "icon-allow-overlap": true,
@@ -68,7 +85,7 @@ const MapLayer: React.FC<MarkerLayerProps> = ({
       };
     }
     addIconImageToMap();
-  }, [map, id, annotationType, source]);
+  }, [map, id, annotationType, source, visibility]);
 
   return null; // No rendering needed, as this component only adds a layer to the map
 };