+++ src/redux/reducers/filtersReducer.ts
@@ -1,58 +1,85 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-export interface IFilterSetting {
-  featureType: string;
-  include: boolean;
+export interface LayerVisibilityState {
+  [layerId: string]: boolean;
+}
+export interface FeaturesVisibilityState {
+  [layerId: string]: boolean;
+}
+export interface IFilterVisibilityState {
+  [filter: string]: {
+    [id: number]: {
+      visible: boolean;
+    };
+  };
 }
 
 export interface IFiltersState {
-  [filterCriteria: string]: Record<number, boolean>; // Use Record<number, boolean> for feature IDs
+  // Use Record<number, boolean> for feature IDs
+  features: IFilterVisibilityState;
+  layerVisibility: LayerVisibilityState;
 }
 
-const initialState: IFiltersState = {};
+const initialState: IFiltersState = { features: {}, layerVisibility: {} };
 
 const filtersSlice = createSlice({
   name: "filters",
   initialState,
   reducers: {
-    setFilterVisibility: (
+    setFilterVisibility(
       state,
       action: PayloadAction<{
         filterCriteria: string;
         featureId: number;
         visible: boolean;
       }>,
-    ) => {
+    ) {
       const { filterCriteria, featureId, visible } = action.payload;
-      // Create the filter criteria object if it doesn't exist
-      if (!state[filterCriteria]) {
-        state[filterCriteria] = {};
-      }
-
-      // Update the visibility for the specified feature type and featureId
-      state[filterCriteria][featureId] = visible;
+      console.log(featureId, visible);
+      return {
+        ...state,
+        features: {
+          ...state.features,
+          [filterCriteria]: {
+            ...state.features[filterCriteria],
+            [featureId]: {
+              visible,
+            },
+          },
+        },
+      };
     },
-    toggleAllLayerAnnotations: (state, action: PayloadAction<string>) => {
-      const filterCriteria = action.payload;
-
-      // Ensure the filter criteria exists in the state
-      if (state[filterCriteria]) {
-        // Iterate through the filter criteria and set all items to 'false'
-        Object.keys(state[filterCriteria]).forEach((featureId) => {
-          state[filterCriteria][featureId] = { visible: false };
+    toggleAllLayerAnnotations: (
+      state,
+      action: PayloadAction<{
+        filterCriteria: string;
+        visible: boolean;
+      }>,
+    ) => {
+      const { filterCriteria, visible } = action.payload;
+      const newState = { ...state };
+      newState.layerVisibility[filterCriteria] = visible;
+      if (state.features[filterCriteria]) {
+        Object.keys(newState.features[filterCriteria]).forEach((featureId) => {
+          newState.features[filterCriteria] = {
+            ...newState.features[filterCriteria], // Preserve existing filters
+            [`${featureId}`]: { visible }, // Update the specific feature
+          };
         });
+        Object.assign(state, newState);
       }
     },
     updateFilterTypes: (
       state,
-      action: PayloadAction<{
-        [filterCriteria: string]: Record<string, { visible: boolean }>;
-      }>,
+      action: PayloadAction<IFilterVisibilityState>,
     ) => {
-      // Merge the updated filter types into the state
+      console.log(action.payload);
       Object.keys(action.payload).forEach((filterCriteria) => {
-        state[filterCriteria] = {
-          ...state[filterCriteria],
+        if (!state.layerVisibility.hasOwnProperty(filterCriteria)) {
+          state.layerVisibility[filterCriteria] = true; // Create and set to true
+        }
+        state.features[filterCriteria] = {
+          ...state.features[filterCriteria],
           ...action.payload[filterCriteria],
         };
       });